+++ test/android/instrumented/src/main/kotlin/me/proton/core/test/android/instrumented/ui/espresso/OnListView.kt
@@ -117,7 +117,7 @@ class OnListView : ConditionWatcher {
 
     /** Builds [DataInteraction] based on parameters provided to [OnListView.Builder]. **/
     private fun dataInteraction(viewAssertion: ViewAssertion = matches(ViewMatchers.isDisplayed())): DataInteraction {
-        waitForCondition({ onData(dataMatcher).check(viewAssertion) })
+        waitForCondition { onData(dataMatcher).check(viewAssertion) }
         return onData(dataMatcher)
     }
 }